requirements.txt
streamlit
streamlit-lottie
scholarly
pandas
graphviz
fuzzywuzzy
python-Levenshtein
apscheduler
email-validator
python-dotenv
requests
beautifulsoup4